+++ core/gsb-api/Cargo.toml
@@ -20,3 +20,4 @@ anyhow = "1"
 log = "0"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
+lazy_static = "1"
